Refactor controllers imports & exports
