+++ src/controllers/v1/index.ts
@@ -1,1 +1,5 @@
-export { MessagesController } from './messagesController'
+import MessagesController from './messagesController'
+
+export {
+  MessagesController,
+}
